*the explaination of the cors lines : 1- config.setAllowedOrigins(Collections.singletonList("*")) : this line is to allow all origins to access the resources 2- config.setAllowedMethods(Collections.singletonList("*")) : this line is to allow all methods to access the resources 3- config.setAllowCredentials(true) : this line is to allow the credentials to be sent to the server 4- config.setAllowedHeaders(Collections.singletonList("*")) : this line is to allow all headers to access the resources 5- config.setMaxAge(3600L) : to set the maximum age of the cache so that the browser does not send the preflight request again
