+++ src/main/java/com/example/cloud_tracker/configuration/CorsConfig.java
@@ -1,0 +1,20 @@
+package com.example.cloud_tracker.configuration;
+
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import jakarta.servlet.http.HttpServletRequest;
+import java.util.Collections;
+
+public class CorsConfig implements CorsConfigurationSource {
+
+    @Override
+    public CorsConfiguration getCorsConfiguration(HttpServletRequest request) {
+        CorsConfiguration config = new CorsConfiguration();
+        config.setAllowedOrigins(Collections.singletonList("*"));
+        config.setAllowedMethods(Collections.singletonList("*"));
+        config.setAllowCredentials(true);
+        config.setAllowedHeaders(Collections.singletonList("*"));
+        config.setMaxAge(3600L);
+        return config;
+    }
+}
